Normalise storage handlers during SuperTokens.init
