+++ lib/ts/recipe/emailverification/recipe.ts
@@ -13,7 +13,7 @@
  * under the License.
  */
 
-import { CreateRecipeFunction, NormalisedAppInfo, StorageHandlerInput } from "../../types";
+import { CreateRecipeFunction, NormalisedAppInfo, NormalisedStorageHandlers } from "../../types";
 import RecipeModule from "../recipeModule";
 import { InputType, NormalisedInputType, PreAndPostAPIHookAction, RecipeInterface } from "./types";
 import { normaliseUserInput } from "./utils";
@@ -37,19 +37,19 @@ export default class Recipe implements RecipeModule<PreAndPostAPIHookAction, Nor
                 appInfo: this.config.appInfo,
                 preAPIHook: this.config.preAPIHook,
                 postAPIHook: this.config.postAPIHook,
-                storageHandlerInput: this.config.storageHandlerInput,
+                storageHandlers: this.config.storageHandlers,
             })
         );
         this.recipeImplementation = builder.override(this.config.override.functions).build();
     }
 
     static init(config?: UserInput): CreateRecipeFunction<PreAndPostAPIHookAction> {
-        return (appInfo: NormalisedAppInfo, storageHandlerInput?: StorageHandlerInput) => {
+        return (appInfo: NormalisedAppInfo, storageHandlers: NormalisedStorageHandlers) => {
             Recipe.instance = new Recipe({
                 ...config,
                 appInfo,
                 recipeId: Recipe.RECIPE_ID,
-                storageHandlerInput,
+                storageHandlers,
             });
 
             return Recipe.instance;
