+++ lib/ts/recipe/recipeModule/types.ts
@@ -12,7 +12,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-import { NormalisedAppInfo, StorageHandlerInput } from "../../types";
+import { NormalisedAppInfo, NormalisedStorageHandlers } from "../../types";
 
 export type RecipePreAPIHookContext<Action> = {
     requestInit: RequestInit;
@@ -52,7 +52,7 @@ export type UserInput<Action> = {
 };
 
 export type RecipeConfig<Action> = {
-    storageHandlerInput?: StorageHandlerInput;
+    storageHandlers: NormalisedStorageHandlers;
     recipeId: string;
     appInfo: NormalisedAppInfo;
 } & UserInput<Action>;
@@ -60,7 +60,7 @@ export type RecipeConfig<Action> = {
 export type NormalisedRecipeConfig<Action> = {
     recipeId: string;
     appInfo: NormalisedAppInfo;
-    storageHandlerInput?: StorageHandlerInput;
+    storageHandlers: NormalisedStorageHandlers;
     preAPIHook: RecipePreAPIHookFunction<Action>;
     postAPIHook: RecipePostAPIHookFunction<Action>;
 };
@@ -85,5 +85,5 @@ export type RecipeImplementationInput<PrePostAPIHookAction> = {
     appInfo: NormalisedAppInfo;
     preAPIHook: RecipePreAPIHookFunction<PrePostAPIHookAction>;
     postAPIHook: RecipePostAPIHookFunction<PrePostAPIHookAction>;
-    storageHandlerInput?: StorageHandlerInput;
+    storageHandlers: NormalisedStorageHandlers;
 };
