+++ lib/ts/supertokens.ts
@@ -13,8 +13,9 @@
  * under the License.
  */
 
+import { normaliseStorageHandlerInput } from "./common/storage/utils";
 import RecipeModule from "./recipe/recipeModule";
-import { NormalisedAppInfo, SuperTokensConfig } from "./types";
+import { NormalisedAppInfo, NormalisedStorageHandlers, SuperTokensConfig } from "./types";
 import { checkForSSRErrorAndAppendIfNeeded, isTest, normaliseInputAppInfoOrThrowError } from "./utils";
 
 export default class SuperTokens {
@@ -27,10 +28,12 @@ export default class SuperTokens {
      * Instance Attributes.
      */
     appInfo: NormalisedAppInfo;
+    storageHandlers: NormalisedStorageHandlers;
     recipeList: RecipeModule<any, any>[] = [];
 
     constructor(config: SuperTokensConfig) {
         this.appInfo = normaliseInputAppInfoOrThrowError(config.appInfo);
+        this.storageHandlers = normaliseStorageHandlerInput(config.storageHandlers);
 
         if (config.recipeList === undefined || config.recipeList.length === 0) {
             throw new Error(
@@ -39,7 +42,7 @@ export default class SuperTokens {
         }
 
         this.recipeList = config.recipeList.map((recipe) => {
-            return recipe(this.appInfo, config.storageHandlers);
+            return recipe(this.appInfo, this.storageHandlers);
         });
     }
